Removed angle constraints from 2D FABRIK, as they were not working.
This was my fault, I didn't test the code very much and assumed it was working. The code works, but only on the very first joint and only if that first joint is never rotated by
other IK modifiers or any other means. All other joins beyond the first joint do not work. I only discovered this while trying to write constrained FABRIK for Twisted IK 2 and
found that the GSoC code was not working in the plugin, so I tested here with this code and confirmed it does not work.

So, FABRIK will just be unconstrained now, unfortunately. If I find a solution to constrained IK for 2D FABRIK, I will document it so it can be potentially implemented
into this system as well.

Also: Fixed the header in physical_bone_2d.h not using the correct year.